test(graphql): Remove explicit cleanup_resources
## Description
These were originally added to try and deal with a deadlock issue, which
was actually due to an issue with `inotify` on macOS, which is fixed by
#19195.

These calls are safe to remove (also note that if the test failed, they
would never be hit, because the assert would cause a panic before they
ran).

## Test plan

Rebase on top of #19195, run the following on macOS:

```
sui$ cargo nextest run       \
  -j 1 -p sui-graphql-rpc    \
  --test e2e_tests           \
  --features pg_integration
```

This used to intermittently hang, but now succeeds.